Переход на Apache License, проверка config файлов

+++ src/qt/FileUpload/createMainWindow/createMainWindow.cpp
@@ -1,3 +1,19 @@
+/***
+ * Copyright 2019 Alexander Pishchulev (https://github.com/BlasterAlex)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include <QDir>
 #include <QFile>
 #include <QHBoxLayout>
@@ -208,17 +224,20 @@ void FileUpload::refreshList() {
 
 // Вызов окна загрузки файлов
 void FileUpload::callWeb() {
-  if (!web) {
+  if (!web) { // если нет открытого окна загрузки
     dwnMenu->setStyleSheet("color: #f54545");
+    if (!webResFileIsValid()) // проверка файла ресурсов
+      dwnMenu->setStyleSheet("color: black");
+    else {
+      page = new WebLoader(this);
+      page->show();
+      page->filesUpload();
+      web = true;
 
-    page = new WebLoader(this);
-    page->show();
-    page->filesUpload();
-    web = true;
-
-    connect(page, SIGNAL(shutdown()), this, SLOT(noWebDialog()));
-    connect(page, SIGNAL(fileDelRequest(QString)), this, SLOT(removeFile(QString)));
-    connect(page, SIGNAL(fileAddRequest(QString)), this, SLOT(addFile(QString)));
+      connect(page, SIGNAL(shutdown()), this, SLOT(noWebDialog()));
+      connect(page, SIGNAL(fileDelRequest(QString)), this, SLOT(removeFile(QString)));
+      connect(page, SIGNAL(fileAddRequest(QString)), this, SLOT(addFile(QString)));
+    }
   }
 }
